requirements.txt
numpy
scipy
matplotlib
seaborn
torch
pyyaml
svgwrite
rich
pyfiglet
pytest
textual
plotext
